[feat] create MainActivity
