+++ app/src/main/java/com/yourssu/unscramble/presentation/MainActivity.kt
@@ -1,46 +1,18 @@
 package com.yourssu.unscramble.presentation
 
 import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.yourssu.unscramble.ui.theme.Yourssu_rookies_unscrambleTheme
+import android.view.LayoutInflater
+import com.yourssu.unscramble.base.BindActivity
+import com.yourssu.unscramble.databinding.ActivityMainBinding
 
-class MainActivity : ComponentActivity() {
+class MainActivity : BindActivity<ActivityMainBinding>() {
+
+    override fun setBinding(layoutInflater: LayoutInflater): ActivityMainBinding {
+        return ActivityMainBinding.inflate(layoutInflater)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContent {
-            Yourssu_rookies_unscrambleTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background,
-                ) {
-                    Greeting("Android")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier,
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    Yourssu_rookies_unscrambleTheme {
-        Greeting("Android")
+        setContentView(binding.root)
     }
 }
